+++ NEWS.md
@@ -12,6 +12,8 @@
 * Fixed issue with `iterations` argument in `get_predicted()` with _brms_
   models.
 
+* Support the *logitr* package: `get_data()`, `find_variables()` and more.
+
 # insight 0.18.5
 
 ## Breaking