+++ NEWS.md
@@ -1,4 +1,14 @@
+# insight 0.18.7
+
+## Bug fixes
+
+* Fixed behaviour of the `at` argument in `get_datagrid()`.
+
 # insight 0.18.6
+
+## New supported models
+
+* Support the *logitr* package: `get_data()`, `find_variables()` and more.
 
 ## Bug fixes
 
@@ -11,8 +21,6 @@
 
 * Fixed issue with `iterations` argument in `get_predicted()` with _brms_
   models.
-
-* Support the *logitr* package: `get_data()`, `find_variables()` and more.
 
 # insight 0.18.5